config: .gitignore: template with default values for components #24239
This toplevel .gitignore covers all possible generated files
from autotools, compiler, linker, etc...

It has been in use for over a year in all xorg modules and has
prevented a good number of patches.

For example, in this module, src/stamp-h1 was not ignored even
if the .gitignore had just been updated. Files with a .patch
extension were not ignored which sometimes cause a patch to be
checked-in git accidentally.

Acked-by: Daniel Stone <daniel@fooishbar.org>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>