config: remove incorrectly implemented visibility compiler option
CFLAGS is a user variable which should never be set by the configuration.
It allows the user to alter the configuration compiler options.

The visibility is only set for GNU compiler, leaving libraries built
with other compilers with the wrong visibility.

All other xorg libraries set visibilty using _X_EXPORT or _X_HIDDEN.

For the time being, all the symbols will have the default visibility
which does not break anything.

Reviewed-by: Jeremy Huddleston <jeremyhu@apple.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>